BioApp Prototype 1
Ugly code, built to test UI and get a feel for the overall structure.
The first button is currently broken and I haven't fixed it as of this
commit, I will on the next.

Very likely to create prototype version two instead of fixing the (MANY)
problems of this build. This is essentially me messing around and
attempting to figure out some things.